Update pyptv.ipynb
not clear what is the optimisation process here, we modify k1, get maybe a better case, but do we need to change x,y,z? I guess we need sometimes to run full calibration (shaking) of all the relevant parameters together to get a better sortgrid and then repeat the story until we get the global minima.