bug in calibration ori name reading and multiplane
1. there is a bug in calibration - the ori file name is read from
the ptv.par and not from calibration.par

2. in multiplane the names were very strict with the plane1_cam and
plane2_cam where the camera id was added at the end of the string.
I suggest to change it to cam#_plane1, cam#_plane2 so the camera id
will be replacing '#' sign and can be everywhere in the name or in the
path, e.g. cal/cam#/plane1